fix: Cast requests_.size() to double
Preventing integer division issue w.r.t. queue_fullness_ratio

Bug: b/409662583
GitOrigin-RevId: c03508a241adbc5aacc5cd63e3cc64f4dee6247b

+++ src/roma/sandbox/dispatcher/dispatcher.cc
@@ -183,7 +183,7 @@ void Dispatcher::ConsumerImpl(int i) {
       {
         absl::MutexLock lock(&mu_);
         response.metrics[kExecutionMetricQueueFullnessRatio] =
-            requests_.size() / max_pending_requests_;
+            static_cast<double>(requests_.size()) / max_pending_requests_;
       }
       response.metrics[kExecutionMetricActiveWorkerRatio] =
           static_cast<double>(num_active_workers_) / workers_.size();
